requirements.txt
Jinja2==2.8
MarkupSafe==0.23
textfsm==0.3.0
requests
python-slugify
